refactored completed status code
